Smart-Environment-Project-TIGER-3D-Point-Cloud-Generation-with-Diffusion-Process/Tiger-Diffusion/requirements_colab.txt
# PyG wheels (torch-2.3.1 + cu121)
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2
torch-geometric==2.5.0
# install with: -f https://data.pyg.org/whl/torch-2.3.1+cu121.html

# 3-D / metrics
open3d==0.18.0
torchmetrics==1.3.2

# science / utils
scikit-learn==1.3.2
numba==0.59.1
matplotlib>=3.8
tqdm>=4.66
opencv-python-headless>=4.9
Pillow>=10.2
Shapely>=2.0
descartes>=1.1.0
fire>=0.6
